comps/router/src/requirements.txt
aiofiles
aiohttp
docarray[full]
docx2txt
fastapi
httpx
kubernetes
langchain
langchain-community
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-sdk
pillow
prometheus-fastapi-instrumentator
pydantic
pypdf
python-dotenv
python-multipart
pyyaml
requests
semantic-router
shortuuid
uvicorn[standard]